+++ Echo/Echo/Echo.vcxproj.filters
@@ -30,13 +30,25 @@
     <ClCompile Include="Echo_Main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Echo_Message.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="Echo_Error.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="Echo_Input.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Echo_Send_Message.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Echo_Frame.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Echo_Main_Help.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Echo_Phone_Book.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Echo_Receive_Message.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -53,13 +65,25 @@
     <ClInclude Include="Echo_Main.h">
       <Filter>Source Files</Filter>
     </ClInclude>
-    <ClInclude Include="Echo_Message.h">
-      <Filter>Source Files</Filter>
-    </ClInclude>
     <ClInclude Include="Echo_Error.h">
       <Filter>Source Files</Filter>
     </ClInclude>
     <ClInclude Include="Echo_Input.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Echo_Send_Message.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Echo_Frame.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Echo_Main_Help.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Echo_Phone_Book.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Echo_Receive_Message.h">
       <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
